+++ src/AdventureWorks.Application/Interfaces/DbContext/IAdventureWorksDbContext.cs
@@ -1,7 +1,7 @@
 ﻿using AdventureWorks.Domain.Entities;
 using Microsoft.EntityFrameworkCore;
 
-namespace AdventureWorks.Application.Interfaces;
+namespace AdventureWorks.Application.Interfaces.DbContext;
 
 public interface IAdventureWorksDbContext
 {
