+++ src/AdventureWorks.Data/Repositories/ProductRepository.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
-using AdventureWorks.Application.Interfaces;
+using AdventureWorks.Application.Interfaces.Repositories;
 using AdventureWorks.Domain.Entities;
 using AdventureWorks.Infrastructure.DbContexts;
 using Microsoft.EntityFrameworkCore;
